refactor: Move to a more scalable project structure
Initially, the project structure was based on type of  where all
controllers, routes, etc were in their own separate directory.

Now, we are moving to a more scalable project structure where all the
controllers, routes, etc of a single entity (for e.g., user) are in one
single place, i.e., we have moved to a structure based on features.