fix: update refresh token generation to make use of class method
The refresh token was being generated inline instead of using an already
existing dedicated method for the same.

The `generateRefreshToken` method has been updated to set jwtid passed
into it instead of setting it inside. This has been done because the
jwtid is also to be stored in DB and hence is better generated in the
transaction itself.