feat: add registerUser controller to handle user registration
This controller is responsible for validating the request body and
passing the data to the designated service to perform further tasks.
This controller then maps the response from the service to the
RegisterResponseDto imported from shared package to adhere the response
data to the API contract and then returns the response.

Note that error handling and response formatting will be handled later.