feat: implement prismaErrorHandler for handling prisma-specific errors
Prisma throws well-defined errors but it isn't feasible to handle them
inside every try-catch of every prisma call given the sheer number of
types of them. So we need a wrapper to handle all these errors in
one place.
The purpose of this wrapper is to handle all necessary prisma errors
and set appropriate message, custom code, and status codes for the same.
The wrapper throws errors and also logs them on site for debugging.
The error middleware handles a proper response for these errors.