feat: add cookie handling for refresh and CSRF tokens on registration
On successful registration, we send the refresh and CSRF tokens as
cookies while access token is sent as part of response body.

We set httpOnly to true for refresh token to revoke JS access in client
side.
However, we don't set this option for CSRF token in order to use double
submit approach. As per this approach, we send the CSRF token as cookie,
and while making state-change requests, the client creates a CSRF token
header and sends it as part of request, which is then validated against
the token sent as cookie. This is done to prevent CSRF attacks.