fix: update token payload content for refresh token
Initially, the refresh token contained the username as well as user role
in its payload which was redundant as they were never required and
could be a security concern as refresh tokens have a long life-span.

This commit reduces the refresh token payload to the bare minimum to
avoid storing crucial information. The token can be identified with its
unique jwtId stored in DB and the user ID as payload.