refactor: restructure JWT payload types and schemas for type safety
We had to type assert the decoded token upon verification even if the
structure is correct. There was a need for more type-safe approach where
we can make sure the structure of decoded payload on verification is
same as the payload provided while signing of token.

To solve this, we make use of Zod schemas and types extended from JWT
provided payload type. We simply create a payload of defined type on
signing, and on verification, we just parse the decoded token against
schema.

This makes sure the structure of decoded payload on verification is
same as the payload provided while signing of token without the need
for type assertion.