fix: improve error handling for refresh token deletion in AuthService
This commit fixes an issue in refresh token process where if the
provided refresh token is not in database, a Not Found error is thrown
with NOT_FOUND error code which is not correct.
The proper error for the same is Unauthorized Error with INVALID_TOKEN
error code, which is what is fixed in this commit.