feat: Add succeedsJwtId to RefreshToken for orphan cleanup
Introduces the `succeedsJwtId` field to the `RefreshToken` model.
This field stores the JWT ID of the refresh token that the current
token is intended to replace.

This mechanism will be used to identify and remove newly generated,
orphaned refresh tokens from the database. Such orphans can occur
if a token refresh request fails or is aborted by the client after
the new token has been persisted but before the client successfully
receives it and the old token is appropriately handled.