feat: notify admins via SSE on new user registration
This commit enhances the user registration process by dispatching a
Server-Sent Event (SSE) to notify administrators in real-time when a
new user is created.

Upon successful registration, the `AuthService` now uses the `SseService`
to multicast a `USER_EVENT` with the reason `USER_CREATED` to all
clients with the `ADMIN` role. This allows admin interfaces to
dynamically update by refreshing a user list without requiring a manual
page reload.