refactor: pass global logger to prismaErrorHandler
The `prismaErrorHandler` utility was recently updated to require a `pino.Logger` instance.

This commit updates the scheduled tasks, `batchUpdateLastActive` and
`clearExpiredTokens`, to pass the global logger to `prismaErrorHandler`.
This is necessary as these background tasks operate outside the request-response
cycle and do not have access to a request-specific child logger.