fix: move generated Prisma client out of src
The `prisma-client-js` generator outputs JavaScript files that are not
intended to be processed by the TypeScript compiler. The previous
configuration placed the output in `src/generated`, causing `tsc` to
incorrectly attempt to compile these files during the build which actually
did not happen and hence the exports file incorrectly referenced the generated
client which did not exist in dist/ causing type errors.

This commit resolves the issue by:
1.  Changing the Prisma schema to output the generated client to a package-level `generated/` directory, outside of `src/`.
2.  Updating the `tsconfig.json` to explicitly exclude this new `generated/` directory from compilation.
3.  Adjusting internal import paths within the `@members-only/database` package to reflect the new location.
4. Adding generated directory to `.gitignore`

This ensures a clean build process where only the source TypeScript files are compiled,
maintaining the correct import path of generated client, hence resolving the misconfiguration.