refactor: generalize lambda creation helper
This commit refactors the Lambda creation helper to make it more generic, allowing it to create both service and non-service Lambdas with the same function.

The `createServiceLambda` method has been renamed to `createLambda`, and its `env` parameter is now optional.
This enables the creation of functions that do not require environment variables.

The existing service Lambda instantiations have been updated to use this new, more flexible helper.