feat: add user activity worker and SQS queue
This commit introduces a new asynchronous processing system for user activity pings.

It provisions the following resources:
- An SQS queue named `user-activity-queue` to buffer incoming activity events.
- A new worker Lambda function, `user-activity-worker`, created using the generic `createLambda` helper.

The Lambda is configured as an event source for the SQS queue and will be triggered with a batch of up to 11 messages or after a 5-minute batching window. This allows for efficient, asynchronous processing of user activity data.