fix: setup cors in sse-service
The cors has been set up again in sse-service.

This is because the API Gateway makes an API call to the EC2 instance running the SSE server on receiving requests.

Additionally, the service lambdas also make an API call to dispatch event to this service.

This makes cors setup necessary to fix cors-errors.